# contracts/openapi/meeting.json
{
  "components": {
    "schemas": {
...
      "ApiResponseMeetingCreatedResponse": {
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/MeetingCreatedResponse"
          },
          "errorCode": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponsePageResponseMeetingSummary": {
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseMeetingSummary"
          },
          "errorCode": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ArtifactResponse": {
        "properties": {
          "artifactId": {
            "format": "uuid",
            "type": "string"
          },
          "artifactState": {
            "type": "string"
          },
          "artifactType": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "retentionUntil": {
            "format": "date-time",
            "type": "string"
          },
          "sizeBytes": {
            "format": "int64",
            "type": "integer"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "GuestInvitee": {
        "properties": {
          "displayName": {
            "maxLength": 160,
            "minLength": 0,
            "type": "string"
          },
          "emailAddress": {
            "format": "email",
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "displayName"
        ],
        "type": "object"
      },
...
      "JsonNode": {},
...
      "MeetingCreatedResponse": {
        "properties": {
          "meeting": {
            "$ref": "#/components/schemas/MeetingDetailResponse"
          },
          "meetingCode": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MeetingDetailResponse": {
        "properties": {
          "accessScope": {
            "type": "string"
          },
          "agenda": {
            "type": "string"
          },
          "aiNotesAvailable": {
            "type": "boolean"
          },
          "allowJoinBeforeHost": {
            "type": "boolean"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactResponse"
            },
            "type": "array"
          },
          "canHost": {
            "type": "boolean"
          },
          "canModerate": {
            "type": "boolean"
          },
          "decisions": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "defaultCameraEnabled": {
            "type": "boolean"
          },
          "defaultMicrophoneEnabled": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "durationMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "endedAt": {
            "format": "date-time",
            "type": "string"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string"
          },
          "followUpActions": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "guestAccessEnabled": {
            "type": "boolean"
          },
          "lifecycleState": {
            "type": "string"
          },
          "meetingCode": {
            "type": "string"
          },
          "meetingId": {
            "format": "uuid",
            "type": "string"
          },
          "organizerName": {
            "type": "string"
          },
          "organizerPersonPublicId": {
            "format": "uuid",
            "type": "string"
          },
          "organizerUserId": {
            "format": "int64",
            "type": "integer"
          },
          "participantRole": {
            "type": "string"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantResponse"
            },
            "type": "array"
          },
          "provider": {
            "type": "string"
          },
          "recordingAvailable": {
            "type": "boolean"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "transcriptAvailable": {
            "type": "boolean"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "waitingRoomEnabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "MeetingSummary": {
        "properties": {
          "accessScope": {
            "type": "string"
          },
          "agenda": {
            "type": "string"
          },
          "allowJoinBeforeHost": {
            "type": "boolean"
          },
          "attendeeCount": {
            "format": "int32",
            "type": "integer"
          },
          "canHost": {
            "type": "boolean"
          },
          "canModerate": {
            "type": "boolean"
          },
          "defaultCameraEnabled": {
            "type": "boolean"
          },
          "defaultMicrophoneEnabled": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "durationMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string"
          },
          "lifecycleState": {
            "type": "string"
          },
          "meetingCode": {
            "type": "string"
          },
          "meetingId": {
            "format": "uuid",
            "type": "string"
          },
          "organizerName": {
            "type": "string"
          },
          "organizerUserId": {
            "format": "int64",
            "type": "integer"
          },
          "participantRole": {
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          },
          "waitingRoomEnabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "PageResponseMeetingSummary": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeetingSummary"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ParticipantResponse": {
        "properties": {
          "admittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "attendanceState": {
            "type": "string"
          },
          "canSelfUnmute": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string"
          },
          "emailAddress": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "joinRequestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          },
          "leftAt": {
            "format": "date-time",
            "type": "string"
          },
          "organizationName": {
            "type": "string"
          },
          "participantId": {
            "format": "uuid",
            "type": "string"
          },
          "participantRole": {
            "type": "string"
          },
          "personPublicId": {
            "format": "uuid",
            "type": "string"
          },
          "unmuteRequestedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ScheduleMeetingRequest": {
        "properties": {
          "accessScope": {
            "enum": [
              "INTERNAL",
              "INVITED",
              "PUBLIC_CODE"
            ],
            "type": "string"
          },
          "agenda": {
            "maxLength": 8000,
            "minLength": 0,
            "type": "string"
          },
          "allowJoinBeforeHost": {
            "type": "boolean"
          },
          "defaultCameraEnabled": {
            "type": "boolean"
          },
          "defaultMicrophoneEnabled": {
            "type": "boolean"
          },
          "description": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string"
          },
          "durationMinutes": {
            "format": "int32",
            "maximum": 1440,
            "minimum": 5,
            "type": "integer"
          },
          "guestAccessEnabled": {
            "type": "boolean"
          },
          "guestInvitees": {
            "items": {
              "$ref": "#/components/schemas/GuestInvitee"
            },
            "maxItems": 100,
            "minItems": 0,
            "type": "array"
          },
          "participantUserIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "maxItems": 200,
            "minItems": 0,
            "type": "array"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "timeZone": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "title": {
            "maxLength": 240,
            "minLength": 0,
            "type": "string"
          },
          "waitingRoomEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "accessScope",
          "startsAt",
          "timeZone",
          "title"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/meetings": {
      "get": {
        "operationId": "meetings",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 30,
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseMeetingSummary"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "video-meeting-controller"
        ]
      },
      "post": {
        "operationId": "schedule",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleMeetingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMeetingCreatedResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "video-meeting-controller"
        ]
      }
    },
...
  }
}